coffee_shop/products/serializers.py
from django.urls import path, include
from django.contrib.auth.models import User
from rest_framework import routers, serializers, viewsets
from rest_framework.serializers import ModelSerializer
from .models import Product


class ProductSerializer(ModelSerializer):
    class Meta:
        model = Product
        fields = [
            "date_create",
            "date_updated",
            "name",
            "description",
            "price",
            "available",
            "photo",
        ]
