requirements.txt
Django==5.2
Pillow==11.1.0
crispy-tailwind==1.0.3
djangorestframework==3.16.0
django-environ==0.12.0
psycopg2-binary==2.9.10